refactor!: swallow up playa.layout entirely
