+++ playa/image.py
@@ -2,7 +2,7 @@ import os
 import os.path
 import struct
 from io import BytesIO
-from typing import BinaryIO, Literal, Tuple
+from typing import TYPE_CHECKING, BinaryIO, Literal, Tuple
 
 from playa.color import (
     LITERAL_DEVICE_CMYK,
@@ -13,13 +13,16 @@ from playa.color import (
 )
 from playa.exceptions import PDFValueError
 from playa.jbig2 import JBIG2StreamReader, JBIG2StreamWriter
-from playa.layout import Item
 from playa.pdftypes import (
     LITERALS_DCT_DECODE,
     LITERALS_FLATE_DECODE,
     LITERALS_JBIG2_DECODE,
     LITERALS_JPX_DECODE,
 )
+
+if TYPE_CHECKING:
+    from playa.page import Item
+
 
 PIL_ERROR_MESSAGE = (
     "Could not import Pillow. This dependency of pdfminer.six is not "
@@ -103,7 +106,7 @@ class ImageWriter:
         if not os.path.exists(self.outdir):
             os.makedirs(self.outdir)
 
-    def export_image(self, image: Item) -> str:
+    def export_image(self, image: "Item") -> str:
         """Save an LTImage to disk"""
         assert (
             image.itype == "image"
@@ -147,7 +150,7 @@ class ImageWriter:
 
         return name
 
-    def _save_jpeg(self, image: Item) -> str:
+    def _save_jpeg(self, image: "Item") -> str:
         """Save a JPEG encoded image"""
         assert (
             image.itype == "image"
@@ -174,7 +177,7 @@ class ImageWriter:
 
         return name
 
-    def _save_jpeg2000(self, image: Item) -> str:
+    def _save_jpeg2000(self, image: "Item") -> str:
         """Save a JPEG 2000 encoded image"""
         assert image.itype == "image" and image.stream is not None
         data = image.stream.get_data()
@@ -195,7 +198,7 @@ class ImageWriter:
             i.save(fp, "JPEG2000")
         return name
 
-    def _save_jbig2(self, image: Item) -> str:
+    def _save_jbig2(self, image: "Item") -> str:
         """Save a JBIG2 encoded image"""
         assert image.itype == "image" and image.stream is not None
         name, path = self._create_unique_image_name(image, ".jb2")
@@ -227,7 +230,7 @@ class ImageWriter:
 
     def _save_bmp(
         self,
-        image: Item,
+        image: "Item",
         width: int,
         height: int,
         bytes_per_line: int,
@@ -245,7 +248,7 @@ class ImageWriter:
                 i += bytes_per_line
         return name
 
-    def _save_bytes(self, image: Item) -> str:
+    def _save_bytes(self, image: "Item") -> str:
         """Save an image without encoding, just bytes"""
         assert (
             image.itype == "image"
@@ -283,7 +286,7 @@ class ImageWriter:
 
         return name
 
-    def _save_raw(self, image: Item) -> str:
+    def _save_raw(self, image: "Item") -> str:
         """Save an image with unknown encoding"""
         assert (
             image.itype == "image"
@@ -299,7 +302,7 @@ class ImageWriter:
         return name
 
     @staticmethod
-    def _is_jbig2_iamge(image: Item) -> bool:
+    def _is_jbig2_iamge(image: "Item") -> bool:
         assert image.itype == "image" and image.stream is not None
         filters = image.stream.get_filters()
         for filter_name, params in filters:
@@ -307,7 +310,7 @@ class ImageWriter:
                 return True
         return False
 
-    def _create_unique_image_name(self, image: Item, ext: str) -> Tuple[str, str]:
+    def _create_unique_image_name(self, image: "Item", ext: str) -> Tuple[str, str]:
         assert image.itype == "image" and image.name is not None
         name = image.name + ext
         path = os.path.join(self.outdir, name)
